+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Changed
 - [#58](https://github.com/CallistoLabsNYC/samsa/issues/58) ConsumerGroup cannot stream when caller only has a reference to a wrapper struct
+- [#34](https://github.com/CallistoLabsNYC/samsa/issues/34) Add documentation and comments for TopicAssignments & TopicPartition structs
 
 ### Fixed
 - [#56](https://github.com/CallistoLabsNYC/samsa/issues/56) Static Analysis and Build CI actions run with no purpose during tag release