+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Samsa Changelog
 
-## [Unreleased]
+## [0.1.5] - 2024-05-11
 ### Changed
 - [#58](https://github.com/CallistoLabsNYC/samsa/issues/58) ConsumerGroup cannot stream when caller only has a reference to a wrapper struct
 - [#47](https://github.com/CallistoLabsNYC/samsa/issues/47) Examples - Publisher and Consumer do not use same topic.